% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignClimateRecordByZonesToDataframe.R
\name{assignClimateRecordByZonesToDataframe}
\alias{assignClimateRecordByZonesToDataframe}
\title{Assign climate record to zones}
\usage{
assignClimateRecordByZonesToDataframe(zonesShapefile, zoneField,
  mosaicDirectory, variables, years)
}
\arguments{
\item{zonesShapefile}{A SpatialPolygonsDataFrame of the zones which will be assigned climate records. The object should have a unique ID column and be in the WGS geographic coordinate system (same as the Daymet NetCDF files, proj4string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0").}

\item{zoneField}{Character string of the field name describing the unique ID values that define the zones.}

\item{mosaicDirectory}{Character string of the file path to the folder containing the netCDF mosaic files.}

\item{variables}{Vector of character strings indicating the variables to process.}

\item{years}{Vector of numeric values indicating the years to process.}
}
\description{
\code{assignClimateRecordByZonesToDataframe} Spatially averages climate records from the Daymet netCDF mosaic files over zones defined by the shapefile and returns a dataframe. If the shapefile is large enough to cause memory problems, the function can be iterated over the results of the "tileShapefile" function.
}
\examples{
assignClimateRecordByZonesToDataframe(spatialPolygonsDataFrame,
                                       "UNIQUE_ID",
                                       "C:/USER/Data/Daymet",
                                       c("tmin", "tmax", "prcp"),
                                       1980:1990)
}

