% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignClimateRecordToZones.R
\name{assignClimateRecordToZones}
\alias{assignClimateRecordToZones}
\title{Assign climate record to zones}
\usage{
assignClimateRecordToZones(zonesShapefile, zoneField, daymetDirectory,
  variables, years, outputToSQLite = FALSE, databaseFilePath = NULL,
  databaseTableName = NULL)
}
\arguments{
\item{zonesShapefile}{A SpatialPolygonsDataFrame of the zones which will be assigned climate records. The object should be in the WGS geographic coordinate system (same as the Daymet NetCDF files, proj4string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0").}

\item{zoneField}{Character string of the field name describing the values that define the zones.}

\item{daymetDirectory}{Character string of the file path to the folder containing the Daymet netCDF files.}

\item{variables}{vector of character strings, the variables to process.}

\item{years}{numeric, a vector of years to process.}

\item{outputToSQLite}{logical; if TRUE the spatially averaged climate records for the zones are output into a SQLite database. In this case, "databaseDirectory" and "databaseTableName" must be defined. If FALSE, the output is returned as a dataframe.}

\item{databaseFilePath}{Character string of the file path to the database to write to. If it does not exist, one will be created.}

\item{databaseTableName}{Character string of the name of the table to create/write to in the database.}
}
\description{
\code{assignClimateRecordToZones} Spatially averages climate records from the Daymet NetCDFs over zones defined by the shapefile.
}

