% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignZonalRecordsToDataframe.R
\name{assignZonalRecordsToDataframe}
\alias{assignZonalRecordsToDataframe}
\title{Assign climate record to zones}
\usage{
assignZonalRecordsToDataframe(zonesShapefile, zoneField, mosaicDirectory,
  variables, years)
}
\arguments{
\item{zonesShapefile}{A SpatialPolygonsDataFrame of the zones which will be assigned climate records.
The object should have a unique ID column and be in the WGS geographic coordinate system (same as the
Daymet NetCDF files, proj4string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0").}

\item{zoneField}{Character string of the field name describing the unique ID values that define the zones.}

\item{mosaicDirectory}{Character string of the file path to the folder containing the netCDF mosaic files.}

\item{variables}{Vector of character strings indicating the variables to process.}

\item{years}{Vector of numeric values indicating the years to process.}
}
\description{
The \code{assignZonalRecordsToDataframe} function patially averages climate records from the
netCDF mosaic files over zones defined by spatial polygons and returns a dataframe.
}
\details{
The netCDF files are read and spatially indexed using custom internal functions. The climate time series
are assigned to the zones represented by the "zonesShapefile" object. If only one record falls into a polygon
it is assigned directly. If multuple records fall into a single polygon, they are averaged. If no records
fall into a polygon, the record nearest to the polygon centroid is assigned. The ouput is written to a
dataframe in long format with columns for the "zoneField", date, and each of the climate variables.

The function relies on the default naming scheme for the netCDF mosaics (e.g. "prcp_2008.nc4").

If the original shapefile is large enough to cause memory problems, the function can be iterated over the
results of the \code{tileShapefile} function which splits the spatial polygon into manageable chunks.
}
\examples{
assignZonalRecordsToDataframe(zonesShapefile  = spatialPolygonsDataFrame,
                              zoneField       = "UNIQUE_ID",
                              mosaicDirectory = "C:/USER/Data/Daymet",
                              variables       = c("tmin", "tmax", "prcp"),
                              years           = 1980:1990)
}

