% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/downloadDaymetMosaic.R
\name{downloadDaymetMosaic}
\alias{downloadDaymetMosaic}
\title{Download Daymet Mosaic NetCDF files.}
\usage{
downloadDaymetMosaic(years, variables, destinationFolder, retryFailedDownloads)
}
\arguments{
\item{years}{A numeric vector specifying the years to download}

\item{variables}{A character vector specifying the variables to download}

\item{destinationFolder}{Character string of the file path to the target folder for downloaded files}

\item{retryFailedDownloads}{logical; If TRUE the function will check (by opening) to see if the files are corrupt.}
}
\description{
\code{downloadDaymetMosaic} downloads mosaic NetCDF files from the Daymet THREDDS server for the variables and years specified.
}

