% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadMosaic.R
\name{downloadMosaic}
\alias{downloadMosaic}
\title{Download netCDF mosaic files.}
\usage{
downloadMosaic(years, variables, destinationFolder,
  retryFailedDownloads = TRUE)
}
\arguments{
\item{years}{A numeric vector specifying the years to download}

\item{variables}{A character vector specifying the variables to download}

\item{destinationFolder}{Character string of the file path to the target folder for downloaded files}

\item{retryFailedDownloads}{logical; If TRUE the function will check (by opening) to see if the netCDF files are corrupt.}
}
\description{
The \code{downloadMosaic} function downloads mosaic NetCDF files from the Daymet THREDDS server for the variables and years specified.
}
\examples{
downloadMosaic(1980:1990,
               c("tmin", "tmax", "prcp"),
               "C:/USER/Data/Daymet",
               TRUE)
}

