% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/returnRecordsByCoordinates.R
\name{returnRecordsByCoordinates}
\alias{returnRecordsByCoordinates}
\title{Return climate records based on coordinate bounds}
\usage{
returnRecordsByCoordinates(areaExtent, mosaicFile, outputFormat = "long",
  startDate, endDate)
}
\arguments{
\item{areaExtent}{Vector string of numeric values specifying lat/lon coordinates of the bounding
box. The format is: c(Minimum Longitude, Maximum Longitude, Minimum Latitude, Maximum Latitude).}

\item{mosaicFile}{Character string of the file path to the netCDF mosiac file to access.}

\item{outputFormat}{A character string of the format of the dataframe output. Options are "wide"
or "long" defaulting to "long".}

\item{startDate}{A character string of the first date to return. This date must match the year of
the mosaic file. The format is: "yyyy-mm-dd"}

\item{endDate}{A character string of the last date to return. This date must match the year of the
mosaic file. The format is: "yyyy-mm-dd"}
}
\description{
The \code{returnRecordsByCoordinates} function returns the climate records for a section of the
mosaic file specified by coordinate bounds.
}
\details{
Climate records for all points within the bounding box will be retrieved and identified by their
source latitude and longitude coordinates.
}
\examples{
returnRecordsByCoordinates(areaExtent   = c(-71.0, -70.0, 42.0, 43.0),
                           mosaicFile   = "C:/USER/Data/Daymet/prcp_2010.nc4",
                           outputFormat = "long",
                           startDate    = "2010-01-01",
                           endDate      = "2010-01-31")
}

