% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/returnZonalRecordsFromDatabase.R
\name{returnZonalRecordsFromDatabase}
\alias{returnZonalRecordsFromDatabase}
\title{Return climate records from SQLite database.}
\usage{
returnZonalRecordsFromDatabase(databaseFilePath, databaseTableName, zoneField,
  startDate, endDate, zoneIDs, variables)
}
\arguments{
\item{databaseFilePath}{Character string of the file path to the SQLite database to access.}

\item{databaseTableName}{Character string of the name of the table to access in the database.}

\item{zoneField}{Character string of the field name describing the unique ID values that define the zones.}

\item{startDate}{Character string of the first date (format: "yyyy-mm-dd") to acces from the database}

\item{endDate}{Character string of the last date (format: "yyyy-mm-dd") to acces from the database}

\item{zoneIDs}{Vector of unique IDs specifying the zones to return}

\item{variables}{Vector of character strings indicating the variables to return}
}
\description{
The \code{returnZonalRecordsFromDatabase} function returns the climate records for specified zones as a dataframe from a SQLite database.
}
\details{
The function provides an option consistent with the package for returning values from the SQLite database.
The "dplyr" package or SQL code can also be used to access climate records from this database in R.
}
\examples{
returnZonalRecordsFromDatabase(databaseFilePath  = "C:/USER/Data/Databases/DaymetDB",
                               databaseTableName = "climate_record",
                               zoneField         = "UNIQUE_ID",
                               startDate         = "1980-01-01",
                               endDate           = "1980-12-31",
                               zoneIDs           = c(1, 2, 3, 4, 5),
                               variables         = c("tmin", "tmax", "prcp"))
}

