% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialAverageSingleFile.R
\name{spatialAverageSingleFile}
\alias{spatialAverageSingleFile}
\title{Spatially average Daymet cells by zone}
\usage{
spatialAverageSingleFile(daymetMosaicFilePath, spatialIndeces, zoneField)
}
\arguments{
\item{daymetMosaicFilePath}{Character string of the file path to the daymet file}

\item{spatialIndeces}{List of 2 dataframes describing the spatial relationships between Daymet NetCDFs and shapefile provided. This is the direct output from the "determineSpatialRelationships" function.}

\item{zoneField}{Character string of the field name describing the values that define the zones}
}
\description{
Returns the spatial average from a single Daymet file for the zones specified in the "spatialIndeces" input object.
}

