% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assignZonalRecordsToDatabase.R
\name{assignZonalRecordsToDatabase}
\alias{assignZonalRecordsToDatabase}
\title{Assign climate record to zones}
\usage{
assignZonalRecordsToDatabase(zonesShapefile, zoneField, zoneFieldType = NULL,
  mosaicDirectory, variables, years, databaseFilePath = NULL,
  databaseTableName = NULL)
}
\arguments{
\item{zonesShapefile}{A SpatialPolygonsDataFrame of the zones which will be assigned climate records.
The object should have a unique ID column and be in the WGS geographic coordinate system (same as the
Daymet NetCDF files, proj4string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0").}

\item{zoneField}{Character string of the field name describing the unique ID values that define the zones.}

\item{zoneFieldType}{Character string defining the column type for the unique ID field written into the
database. Options are "character", "integer", "numeric", or the default: NULL. A NULL value will base the
type on the shapefile field type as interpreted by R.}

\item{mosaicDirectory}{Character string of the file path to the folder containing the netCDF mosaic files.}

\item{variables}{Vector of character strings indicating the variables to process.}

\item{years}{Vector of numeric values indicating the years to process.}

\item{databaseFilePath}{Character string of the file path to the database to write to. If it does not
exist, one will be created.}

\item{databaseTableName}{Character string of the name of the table to create/write to in the database.}
}
\description{
The \code{assignZonalRecordsToDatabase} function spatially averages climate records from the
netCDF mosaic files over zones defined by spatial polygons and exports directly to a SQLite database.
If the shapefile is large enough to cause memory problems, the function can be iterated over the
results of the {tileShapefile} function which splits the spatial polygon into manageable chunks. The
function relies on the default naming scheme for the netCDF mosaics (e.g. "prcp_2008.nc4").
}
\examples{
assignZonalRecordsToDatabase(zonesShapefile    = spatialPolygonsDataFrame,
                             zoneField         = "UNIQUE_ID",
                             zoneFieldType     = "integer",
                             mosaicDirectory   = "C:/USER/Data/Daymet",
                             variables         = c("tmin", "tmax", "prcp"),
                             years             = 1980:1990,
                             databaseFilePath  = "C:/USER/Data/Databases/DaymetDB",
                             databaseTableName = "climate_record")
}

