% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/determineSpatialRelationships.R
\name{determineSpatialRelationships}
\alias{determineSpatialRelationships}
\title{Determine spatial relationships between NetCDF and zone shapefile}
\usage{
determineSpatialRelationships(zonesShapefile, zoneField, exampleDaymetFile)
}
\arguments{
\item{zonesShapefile}{SpatialPolygonsDataFrame A shapefile defining the zones that are to be assigned Daymet records. This needs to be in the same coordinate system(WGS) as the Daymet points.}

\item{zoneField}{Character string of the field name describing the values that define the zones}

\item{exampleDaymetFile}{Character string of the file path to the daymet file used to define the spatial relationships between the shapefile and the Daymet cells}
}
\description{
Defines the spatial relationships between the Daymet NetCDF files and the zones shapefile
}

