% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadMosaic.R
\name{downloadMosaic}
\alias{downloadMosaic}
\title{Download netCDF mosaic files.}
\usage{
downloadMosaic(years, variables, destinationDirectory,
  retryFailedDownloads = TRUE)
}
\arguments{
\item{years}{A numeric vector specifying the years to download}

\item{variables}{A character vector specifying the variables to download}

\item{destinationDirectory}{Character string of the file path to the target folder for downloaded files}

\item{retryFailedDownloads}{logical; If TRUE the function will check (by opening) to see if the netCDF files are corrupt.}
}
\description{
The \code{downloadMosaic} function downloads mosaic NetCDF files from the Daymet THREDDS server
for the variables and years specified. Currently, the web address is hard-coded into the function.
If and when this address is updated, a change will need to be made to the function code.
}
\examples{
downloadMosaic(years                = 1980:1990,
               variables            = c("tmin", "tmax", "prcp"),
               destinationDirectory = "C:/USER/Data/Daymet",
               retryFailedDownloads = TRUE)
}

