% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tileShapefile.R
\name{tileShapefile}
\alias{tileShapefile}
\title{Break shapefile into tiles}
\usage{
tileShapefile(shapefile)
}
\arguments{
\item{shapefile}{A SpatialPolygonsDataFrame of the zones which will be assigned climate records. The object should be in the WGS geographic coordinate system (same as the Daymet NetCDF files,The proj4string = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0").}
}
\description{
\code{tileShapefile} breaks up large shapefiles into up into 2 degree tiles. to keep R from having memory issues when reading Daymet records.
}
\details{
Polygons are assigned to tiles based on where their centroids are located.
This function only works for shapefiles with units that are lat/lon. Make sure the shapefile is in the daymet coordinate system before running this function on it.
}

